Ensure Pids are associated with an alive herp_client.
This removes the race-condition when a herp_client died and the
reference was attempted to be used before the new herp_client was
started. We do this by looping over is_process_alive and waiting until
it is. A macro is defined for the max number of timeouts allowed when
waiting for a reference to 'become' valid.